requirements.txt
PyJWT==2.8.0
cryptography==41.0.7
fastapi==0.104.1
httpx==0.24.1
pydantic-settings==2.1.0
pydantic==2.5.2
pytest==7.4.3
uvicorn==0.24.0.post1
